python/pyproject.toml
[tool.poetry]
name = "grpc-rendezvous"
version = "0.1.0"
description = ""
authors = ["Nick Cao <nickcao@nichi.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.dependencies.jumpstarter-protocol]
git = "https://github.com/jumpstarter-dev/jumpstarter-protocol.git"
branch = "wip"
subdirectory = "python"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
